Move TZE components to a separate module & add TzeOutPoint
