+++ zcash_primitives/src/transaction/mod.rs
@@ -508,7 +508,7 @@ pub mod testing {
     };
 
     #[cfg(feature = "zfuture")]
-    use super::components::{TzeIn, TzeOut};
+    use super::components::{TzeIn, TzeOut, TzeOutPoint};
 
     pub const VALID_OPCODES: [u8; 8] = [
         0x00, // OP_FALSE,
@@ -553,6 +553,13 @@ pub mod testing {
 
     #[cfg(feature = "zfuture")]
     prop_compose! {
+        pub fn arb_tzeoutpoint()(hash in prop::array::uniform32(1u8..), n in 1..100u32) -> TzeOutPoint {
+            TzeOutPoint::new(hash, n)
+        }
+    }
+
+    #[cfg(feature = "zfuture")]
+    prop_compose! {
         pub fn arb_witness()(extension_id in 0..100u32, mode in 0..100u32, payload in vec(any::<u8>(), 32..256))  -> tze::Witness {
             tze::Witness { extension_id, mode, payload }
         }
@@ -560,7 +567,7 @@ pub mod testing {
 
     #[cfg(feature = "zfuture")]
     prop_compose! {
-        pub fn arb_tzein()(prevout in arb_outpoint(), witness in arb_witness()) -> TzeIn {
+        pub fn arb_tzein()(prevout in arb_tzeoutpoint(), witness in arb_witness()) -> TzeIn {
             TzeIn { prevout, witness }
         }
     }
